pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlserver-pg-cdc"
version = "0.1.0"
description = "SQL Server to PostgreSQL CDC Pipeline with reconciliation tooling"
readme = "CLAUDE.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CDC Pipeline Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database :: Database Engines/Servers",
]

dependencies = [
    "psycopg2-binary>=2.9.9",
    "pyodbc>=5.0.1",
    "hvac>=2.1.0",
    "APScheduler>=3.10.4",
    "prometheus-client>=0.19.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.21.0",
    "opentelemetry-instrumentation-psycopg2>=0.42b0",
    "opentelemetry-instrumentation-requests>=0.42b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "testcontainers>=3.7.1",
    "black>=23.12.1",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "hypothesis>=6.92.0",
    "mutmut>=2.4.4",
    "locust>=2.20.0",
    "pact-python>=2.2.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
reconcile = "reconciliation.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
    "--hypothesis-show-statistics",
]
markers = [
    "e2e: End-to-end tests requiring real database instances",
    "vault: Tests requiring HashiCorp Vault",
    "slow: Tests that take significant time to run",
    "integration: Integration tests",
    "contract: Contract tests",
    "performance: Performance tests",
    "property: Property-based tests using Hypothesis",
    "load: Load and stress tests",
]
# Ignore e2e tests for coverage calculation since they don't execute source code
# when databases aren't available
norecursedirs = [".*", "build", "dist", "*.egg"]

[tool.hypothesis]
max_examples = 100
deadline = 1000
derandomize = false

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.mutmut]
paths_to_mutate = "src/reconciliation"
runner = ".venv/bin/pytest -x --tb=line -q --no-cov tests/unit/ tests/property/"
tests_dir = "tests/"
use_coverage = true
test_time_multiplier = 1.5
